--- a/scripts/parse-bidi-brackets.js
+++ b/scripts/parse-bidi-brackets.js
@@ -1,7 +1,6 @@
 'use strict';
 
-const looseMatch = require('unicode-loose-match');
-const propertyAliases = require('unicode-property-aliases');
+const regenerate = require('regenerate');
 const utils = require('./utils.js');
 
 const bidiBracketMap = new Map([
@@ -11,19 +10,22 @@ const bidiBracketMap = new Map([
 ]);
 
 const parseBidiBrackets = function(version) {
-	const map = {};
+	const map = {
+		'Open': regenerate(),
+		'Close': regenerate(),
+		'None': regenerate().addRange(0, 0x10FFFF)
+	};
 	const source = utils.readDataFile(version, 'bidi-brackets');
 	if (!source) {
 		return;
 	}
 	const lines = source.split('\n');
-	const symbolsHandled = new Set();
-	lines.forEach(function(line) {
+	for (const line of lines) {
 		if (
 			/^#/.test(line) ||
 			!/;\x20/.test(line)
 		) {
-			return;
+			continue;
 		}
 		const data = line.trim().split(';');
 		const charRange = data[0].replace('..', '-').trim();
@@ -31,26 +33,19 @@ const parseBidiBrackets = function(version) {
 		item = bidiBracketMap.get(item);
 		const rangeParts = charRange.split('-');
 		if (rangeParts.length == 2) {
-			utils.range(
+			const [from, to] = [
 				parseInt(rangeParts[0], 16),
-				parseInt(rangeParts[1], 16)
-			).forEach(function(codePoint) {
-				symbolsHandled.add(codePoint);
-				utils.append(map, item, codePoint);
-			});
+				parseInt(rangeParts[1], 16),
+			];
+			map['None'].removeRange(from, to);
+			map[item].addRange(from, to);
 		} else {
 			const codePoint = parseInt(charRange, 16);
-			symbolsHandled.add(codePoint);
-			utils.append(map, item, codePoint);
-		}
-	});
-	for (let codePoint = 0x000000; codePoint <= 0x10FFFF; codePoint++) {
-		// Note: `Any`, `ASCII`, and `Assigned` are actually properties,
-		// not categories. http://unicode.org/reports/tr18/#Categories
-		if (!symbolsHandled.has(codePoint)) {
-			utils.append(map, 'None', codePoint);
+			map['None'].remove(codePoint);
+			map[item].add(codePoint);
 		}
 	}
+
 	return map;
 };