--- a/scripts/parse-composition-exclusions.js
+++ b/scripts/parse-composition-exclusions.js
@@ -1,35 +1,32 @@
 'use strict';
 
+const regenerate = require('regenerate');
 const utils = require('./utils.js');
 
 const parseCompositionExclusions = function(version) {
-	const map = {};
 	const source = utils.readDataFile(version, 'composition-exclusions');
 	if (!source) {
 		return;
 	}
+	const result = regenerate();
 	const lines = source.split('\n');
-	lines.forEach(function(line) {
-		if (!line || /^#/.test(line)) {
-			return;
+	for (const line of lines) {
+		if (!line || line.startsWith('#')) {
+			continue;
 		}
 		const data = line.trim().split('#');
-		const propertyName = 'Composition_Exclusion';
 		const charRange = data[0].replace('..', '-').trim();
 		const rangeParts = charRange.split('-');
 		if (rangeParts.length == 2) {
-			utils.range(
-				parseInt(rangeParts[0], 16),
-				parseInt(rangeParts[1], 16)
-			).forEach(function(codePoint) {
-				utils.append(map, propertyName, codePoint);
-			});
+			result.addRange(parseInt(rangeParts[0], 16), parseInt(rangeParts[1], 16));
 		} else {
 			const codePoint = parseInt(charRange, 16);
-			utils.append(map, propertyName, codePoint);
+			result.add(codePoint);
 		}
-	});
-	return map;
+	};
+	return {
+		Composition_Exclusion: result
+	};
 };
 
 module.exports = parseCompositionExclusions;