… data/10.0.0-derived-binary-properties.txt
… data/10.0.0-derived-general-category.txt
… data/11.0.0-derived-binary-properties.txt
… data/11.0.0-derived-general-category.txt
… data/12.0.0-derived-binary-properties.txt
… data/12.0.0-derived-general-category.txt
… data/12.1.0-derived-binary-properties.txt
… data/12.1.0-derived-general-category.txt
… data/13.0.0-derived-binary-properties.txt
… data/13.0.0-derived-general-category.txt
… data/14.0.0-derived-binary-properties.txt
… data/14.0.0-derived-general-category.txt
… data/15.0.0-derived-binary-properties.txt
… data/15.0.0-derived-general-category.txt
… data/15.1.0-derived-binary-properties.txt
… data/15.1.0-derived-general-category.txt
… data/16.0.0-derived-binary-properties.txt
… data/16.0.0-derived-general-category.txt
… data/3.1.0-derived-general-category.txt
… data/3.2.0-derived-binary-properties.txt
… data/3.2.0-derived-general-category.txt
… data/4.0.0-derived-binary-properties.txt
… data/4.0.0-derived-general-category.txt
… data/4.0.1-derived-binary-properties.txt
… data/4.0.1-derived-general-category.txt
… data/4.1.0-derived-binary-properties.txt
… data/4.1.0-derived-general-category.txt
… data/5.0.0-derived-binary-properties.txt
… data/5.0.0-derived-general-category.txt
… data/5.1.0-derived-binary-properties.txt