This PR is a massive rewrite of current build script. The performance improvements come from:

1) Use regenerate when parsing most sources, and avoid expanding into arrays
2) Use derived UCD data which directly provides ranges, doing so can avoid regenerate merge steps.
3) Generate run-length directly from regenerate internal data

Runtime measured in my local environment (Apple M1 Max 10c)
```
# Current main:
$ time npm run build
npm run build  167.54s user 8.65s system 396% cpu 44.384 total

# This PR:
$ time npm run build
npm run build  15.54s user 10.48s system 343% cpu 7.585 total
```

<details>
<summary>The time spent on each task when processing Unicode 16.0.0 in this PR (Click to unfold)</summary>

```
Generating data for Unicode v16.0.0…
Parsing Unicode v16.0.0 `Bidi_Class`
bidi_class: 60.228ms
Parsing Unicode v16.0.0 `Script`…
scripts: 46.526ms
Parsing Unicode v16.0.0 `Script_Extensions`…
script extensions: 48.224ms
Parsing Unicode v16.0.0 properties…
properties: 13.021ms
Parsing Unicode v16.0.0 derived core properties…
derived properties: 35.659ms
Parsing Unicode v16.0.0 derived general category…
derived general category: 31.849ms
Parsing Unicode v16.0.0 derived normalization properties…
derived normalization properties: 54.556ms
Parsing Unicode v16.0.0 derived binary properties…
derived binary properties: 0.646ms
Parsing Unicode v16.0.0 composition exclusions…
compostion exclusions: 0.476ms
Parsing Unicode v16.0.0 `Case_Folding`…
case folding: 8.896ms
Parsing Unicode v16.0.0 `Block`…
block: 83.674ms
Parsing Unicode v16.0.0 `Bidi_Mirroring_Glyph`
bidi mirroring glyph: 230.303ms
Parsing Unicode v16.0.0 bidi brackets…
bidi brackets: 1.76ms
Parsing Unicode v16.0.0 `Line_Break`…
line break: 21.328ms
Parsing Unicode v16.0.0 `Word_Break`…
word break: 9.971ms
Parsing Unicode v16.0.0 binary emoji properties…
binary emoji properties: 5.935ms
Parsing Unicode v16.0.0 emoji sequence properties…
emoji sequence properties: 35.444ms
Parsing Unicode v16.0.0 `Names`…
names: 146.272ms
Parsing Unicode v16.0.0 Aliases…
aliases: 2.588ms
Parsing Unicode v16.0.0 simple case mappings…
simple case mappings: 41.472ms
Parsing Unicode v16.0.0 `Special_Casing`…
special casing: 6.741ms
```
</details>

Since the output is not tracked in the git repository, here I attached the output diff between current main and this PR:
[output.diff.gz](https://github.com/user-attachments/files/17638397/output.diff.gz)

The only output differences are:

1) Fixed incorrect Word_Break/Other and Line_Break/Unknown (See comment below)
2) The `General_Category` property is now defined before the `Binary_Property` property in the `index.js` exports, so as the ordering in `README.md`
3) The encoding utitlies are removed from `./decode-ranges.js`

Closes #3.